Final version of the app
The final version of the app was completed. The report feature was
completed, which shows the user a history of all the items that s/he has
bought. A help page was also added to the app so that the user can find
help there. Validation was also added when saving a new list name. If no
items are selected in the list, if the list name is already taken or if
the list name field is left empty the user will be shown an error
message informing him/her of what needs to be fixed.